Deprecate separate logging config file
Remove separate logging file functionality as it's not generally used
and causes issues when accidentally included. Add a warning, so that if
the user provides a path for the logging config file (or a file
exists at the old delfault path), the user is warned that the
functionality has been deprecated.